Add hide_profile as a standalone UserOption
This commit adds a new user option that is responsible for hiding a
user's profile. It is separate from the `hide_profile_and_presence`
option. That option will be replaced entirely in upcoming commits.

This change is the first in a series of changes that will allow a user
to independently adjust hiding their profile and hiding their presence.

The commit adds the `UserOption#hide_profile` attribute and propagates
that up to stack to the UI, Site Settings, and import tasks.