FIX: Fix flakey private_message_map spec
The participant/user divs are rendered unordered. This made the test fail
sometimes when the order of the user divs didn't match the test, since
`have_link` doesn't search through all the <div> in the list.

The resolution was order the expected users and the user divs. I could
have added an "order by" to view, but I didn't want to create new
behavior to _could_ break some other code's expectations.